Cast current_wordsize to an int
On 64bit systems with a single personality, we see:
count.c: In function 'call_summary':
count.c:223:5: warning: format '%u' expects type 'unsigned int',
	but argument 3 has type 'long unsigned int'

Since on multi-personality systems this is an array of ints, cast
the multiplication to an int and update the printf format.

* count.c (call_summary): Change %u to %d and cast first argument to int.

+++ count.c
@@ -219,8 +219,8 @@ call_summary(FILE *outf)
 			set_personality(i);
 		if (i)
 			fprintf(outf,
-				"System call usage summary for %u bit mode:\n",
-				current_wordsize * 8);
+				"System call usage summary for %d bit mode:\n",
+				(int) (current_wordsize * 8));
 		call_summary_pers(outf);
 	}
